CONCF-591 rename files
